requirements.txt
git+https://github.com/farihdzkyy/deezloadclone
git+https://github.com/kokarare1212/librespot-python
pyrogram
yt-dlp
pymongo
shazamio
aiohttp
aiofiles
youtube-search-python
requests
validators
isodate
pyshorteners
spotipy
py-deezer
deezer-python
pathlib
pydub
pillow
moviepy
azapi
tiktok-downloader
